+++ app/frontend/players/PlayersData.ts
@@ -30,6 +30,38 @@ declare const Routes: {
   ) => string;
 };
 
+export async function loadPlayer(tournamentId: number, playerId: number) {
+  try {
+    const response = await fetch(
+      `/beta/tournaments/${tournamentId}/players/${playerId}`,
+      {
+        method: "GET",
+      },
+    );
+
+    return  (await response.json()) as Player;
+  } catch {
+    globalMessages.errors.push(`Error loading player data for player ${playerId}`);
+    return null;
+  }
+}
+
+export async function loadPlayerByUserId(tournamentId: number, userId: number) {
+  try {
+    const response = await fetch(
+      `/beta/tournaments/${tournamentId}/players/by_user_id/${userId}`,
+      {
+        method: "GET",
+      },
+    );
+
+    return (await response.json()) as Player;
+  } catch {
+    globalMessages.errors.push(`Error loading player data for user ${userId}.`);
+    return null;
+  }
+}
+
 export async function loadPlayers(tournamentId: number) {
   const response = await fetch(
     Routes.players_data_beta_tournament_players_path(tournamentId),
