+++ app/frontend/tests/Tournament.svelte-test.ts
@@ -10,18 +10,16 @@ import Tournament from "../tournaments/Tournament.svelte";
 import { MockIdentityNames, MockTournament } from "./TournamentTestData";
 import { MockPlayerBob } from "./RoundsTestData";
 import {
-  loadPlayer,
   loadQRCode,
   loadTournament,
 } from "../tournaments/TournamentSettings";
 import { loadIdentityNames } from "../identities/Identity";
 import userEvent from "@testing-library/user-event";
-import { Player, savePlayer } from "../players/PlayersData";
+import { loadPlayerByUserId, Player, savePlayer } from "../players/PlayersData";
 
 vi.mock("../tournaments/TournamentSettings", async (importOriginal) => ({
   ...(await importOriginal<typeof import("../pairings/PairingsData")>()),
   loadTournament: vi.fn(() => MockTournament),
-  loadPlayer: vi.fn(() => MockPlayerBob),
   loadQRCode: vi.fn(() => null),
 }));
 
@@ -32,6 +30,7 @@ vi.mock("../identities/Identity", async (importOriginal) => ({
 
 vi.mock("../players/PlayersData", async (importOriginal) => ({
   ...(await importOriginal<typeof import("../pairings/PairingsData")>()),
+  loadPlayerByUserId: vi.fn(() => MockPlayerBob),
   savePlayer: vi.fn(() => true),
 }));
 
@@ -45,7 +44,7 @@ describe("Tournament", () => {
       expect(loadTournament).toHaveBeenCalledOnce();
     });
     await waitFor(() => {
-      expect(loadPlayer).toHaveBeenCalledOnce();
+      expect(loadPlayerByUserId).toHaveBeenCalledOnce();
     });
     await waitFor(() => {
       expect(loadQRCode).toHaveBeenCalledOnce();
@@ -75,7 +74,7 @@ describe("Tournament", () => {
 
   describe("when player is logged out", () => {
     beforeEach(() => {
-      vi.mocked(loadPlayer).mockImplementation(() =>
+      vi.mocked(loadPlayerByUserId).mockImplementation(() =>
         Promise.resolve(new Player()),
       );
     });
@@ -143,7 +142,7 @@ describe("Tournament", () => {
   describe("when player is logged in", () => {
     describe("when player is not registered", () => {
       beforeEach(async () => {
-        vi.mocked(loadPlayer).mockImplementation(() =>
+        vi.mocked(loadPlayerByUserId).mockImplementation(() =>
           Promise.resolve(new Player()),
         );
